合并所有修改代码

+++ src/task/src/Annotation/Task.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
+
 namespace Hyperf\Task\Annotation;
 
 use Attribute;
@@ -17,7 +18,7 @@ use Hyperf\Di\Annotation\AbstractAnnotation;
 #[Attribute(Attribute::TARGET_CLASS | Attribute::TARGET_METHOD)]
 class Task extends AbstractAnnotation
 {
-    public function __construct(public int $workerId = -1, public int $timeout = 10)
+    public function __construct(public int $workerId = -1, public float $timeout = 10)
     {
     }
 }
